Improve logic on how fetched data is rendered in options
Previously data needed to be filtered and sorted correctly on both the query and the component, or otherwise newly fetched data could end up in the middle of the existing options. This makes the logic simpler by just rendering the pages in order they come in the queries. The previous logic had also many anti-patterns and it was about to break on any changes.